requirements.txt
Django==1.9.7
django-appconf==1.0.2
django-bootstrap-form==3.2.1
django-crispy-forms==1.6.0
django-filter==0.14.0
django-ratelimit==1.0.0
django-user-accounts==1.3.1
djangorestframework==3.4.6
pinax-theme-bootstrap==7.10.2
pytz==2016.6.1
wheel==0.24.0
